requirements.txt
streamlit>=1.28.0
youtube-transcript-api>=0.6.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-postgres>=0.0.6
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
openai>=1.0.0
pgvector>=0.2.4